tests/requirements.txt
pytest==8.4.2
requests==2.32.5
fastapi==0.116.1
httpx==0.28.1
pillow==11.3.0
